requirements.txt
fastapi
uvicorn[standard]
httpx
feedparser
anthropic
apscheduler
python-dotenv
